ENH: da.ufuncs now supports DataFrame/Series
