+++ dask/array/__init__.py
@@ -7,14 +7,15 @@ from .core import (Array, stack, concatenate, take, tensordot, transpose,
         digitize, histogram, map_blocks, atop, to_hdf5, dot, cov, array,
         dstack, vstack, hstack, to_npy_stack, from_npy_stack, compress,
         from_delayed, round, swapaxes, repeat)
-from .core import (logaddexp, logaddexp2, conj, exp, log, log2, log10, log1p,
+from .core import (around, isnull, notnull, isclose, eye, triu,
+                   tril, diag, corrcoef)
+from .ufunc import (logaddexp, logaddexp2, conj, exp, log, log2, log10, log1p,
         expm1, sqrt, square, sin, cos, tan, arcsin, arccos, arctan, arctan2,
         hypot, sinh, cosh, tanh, arcsinh, arccosh, arctanh, deg2rad, rad2deg,
         logical_and, logical_or, logical_xor, logical_not, maximum, minimum,
         fmax, fmin, isreal, iscomplex, isfinite, isinf, isnan, signbit,
         copysign, nextafter, ldexp, fmod, floor, ceil, trunc, degrees, radians,
-        rint, fix, angle, real, imag, clip, fabs, sign, absolute, frexp, modf,
-        around, isnull, notnull, isclose, eye, triu, tril, diag, corrcoef)
+        rint, fix, angle, real, imag, clip, fabs, sign, absolute, frexp, modf)
 from .reductions import (sum, prod, mean, std, var, any, all, min, max, vnorm,
                          moment,
                          argmin, argmax,
