+++ dask/array/reductions.py
@@ -9,7 +9,8 @@ import numpy as np
 from toolz import compose, partition_all, merge, get, accumulate, pluck
 
 from . import chunk
-from .core import _concatenate2, Array, atop, sqrt, lol_tuples
+from .core import _concatenate2, Array, atop, lol_tuples
+from .ufunc import sqrt
 from .numpy_compat import divide
 from ..compatibility import getargspec, builtins
 from ..base import tokenize
